requirements.txt
PyPDF2
sentence-transformers
scikit-learn
networkx
python-igraph
leidenalg
